Fix bug with displaying Comments
The error occurred because you were trying to do:

key = db.Key.from_path('Post', int(post_id), parent=blog_key())
        post = db.get(key)
        comments = Comments.by_id(post)

where Comments.by_id(post), should be the post_id, not the post-object.
Of course, the error won't show up there, since filter actually just
filters it for you, so it returns only an empty query-object.

So in the permalink.html, you tried to run a forloop over it, but a query-object is not iteratable, so it throws an error.

The fix is to either send Comments.by_id(post_id), but that is not the best solution, since it would have to do another check for the post, and we already have the post. So instead, just filter on the post-object itself. It is faster.